Cleaned project 1. removed unused dependency (pom.xml) 2. removed empty line (pom.xml) 3. added .gitignore 4. removed package annotation(transferred all files to package lib) 5. renamed addRoleToDB and deleteRoleToDB methods 6. refactored getByToken method in class UserDaoJdbcImpl
