+++ .gitignore
@@ -1,0 +1,4 @@
+.idea/*
+*.iml
+target/*
+*.log
