+++ src/main/java/mate/academy/internetshop/controller/DeleteOrderController.java
@@ -6,7 +6,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import mate.academy.internetshop.annotation.Inject;
+import mate.academy.internetshop.lib.Inject;
 import mate.academy.internetshop.service.OrderService;
 
 public class DeleteOrderController extends HttpServlet {
