+++ src/main/java/mate/academy/internetshop/controller/InjectInitializer.java
@@ -3,7 +3,7 @@ package mate.academy.internetshop.controller;
 import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
 
-import mate.academy.internetshop.annotation.Injector;
+import mate.academy.internetshop.lib.Injector;
 import org.apache.log4j.Logger;
 
 public class InjectInitializer implements ServletContextListener {
