+++ src/main/java/mate/academy/internetshop/controller/LoginController.java
@@ -8,8 +8,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import mate.academy.internetshop.annotation.Inject;
 import mate.academy.internetshop.exception.AuthenticationException;
+import mate.academy.internetshop.lib.Inject;
 import mate.academy.internetshop.model.User;
 import mate.academy.internetshop.service.BucketService;
 import mate.academy.internetshop.service.UserService;
