+++ src/main/java/mate/academy/internetshop/dao/RoleDao.java
@@ -7,9 +7,9 @@ import mate.academy.internetshop.model.User;
 
 public interface RoleDao {
 
-    Set<Role> addRoleToDB(User user);
+    Set<Role> addRole(User user);
 
-    void deleteRoleToDB(Long userId);
+    void deleteRole(Long userId);
 
     Role getRoleByName(Role.RoleName roleName);
 }
