+++ src/main/java/mate/academy/internetshop/lib/Inject.java
@@ -1,4 +1,4 @@
-package mate.academy.internetshop.annotation;
+package mate.academy.internetshop.lib;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
