+++ src/main/java/mate/academy/internetshop/service/impl/BucketServiceImpl.java
@@ -2,9 +2,9 @@ package mate.academy.internetshop.service.impl;
 
 import java.util.List;
 
-import mate.academy.internetshop.annotation.Inject;
 import mate.academy.internetshop.dao.BucketDao;
 import mate.academy.internetshop.dao.ItemDao;
+import mate.academy.internetshop.lib.Inject;
 import mate.academy.internetshop.lib.Service;
 import mate.academy.internetshop.model.Bucket;
 import mate.academy.internetshop.model.Item;
