+++ src/main/java/mate/academy/internetshop/service/impl/ItemServiceImpl.java
@@ -2,8 +2,8 @@ package mate.academy.internetshop.service.impl;
 
 import java.util.List;
 
-import mate.academy.internetshop.annotation.Inject;
 import mate.academy.internetshop.dao.ItemDao;
+import mate.academy.internetshop.lib.Inject;
 import mate.academy.internetshop.lib.Service;
 import mate.academy.internetshop.model.Item;
 import mate.academy.internetshop.service.ItemService;
