+++ src/main/java/mate/academy/internetshop/service/impl/OrderServiceImpl.java
@@ -3,9 +3,9 @@ package mate.academy.internetshop.service.impl;
 
 import java.util.List;
 
-import mate.academy.internetshop.annotation.Inject;
 import mate.academy.internetshop.dao.OrderDao;
 import mate.academy.internetshop.dao.UserDao;
+import mate.academy.internetshop.lib.Inject;
 import mate.academy.internetshop.lib.Service;
 import mate.academy.internetshop.model.Item;
 import mate.academy.internetshop.model.Order;
