+++ src/main/java/mate/academy/internetshop/service/impl/RoleServiceImpl.java
@@ -1,7 +1,7 @@
 package mate.academy.internetshop.service.impl;
 
-import mate.academy.internetshop.annotation.Inject;
 import mate.academy.internetshop.dao.RoleDao;
+import mate.academy.internetshop.lib.Inject;
 import mate.academy.internetshop.lib.Service;
 import mate.academy.internetshop.model.Role;
 import mate.academy.internetshop.service.RoleService;
